move mock users store and service accounts store
