+++ v2/apiserver/internal/authn/users_test.go
@@ -18,7 +18,7 @@ func TestUserListMarshalJSON(t *testing.T) {
 }
 
 func TestNewUsersService(t *testing.T) {
-	usersStore := &mockUsersStore{}
+	usersStore := &MockUsersStore{}
 	sessionsStore := &mockSessionsStore{}
 	svc := NewUsersService(usersStore, sessionsStore)
 	require.Same(t, usersStore, svc.(*usersService).usersStore)
@@ -34,7 +34,7 @@ func TestUserServiceList(t *testing.T) {
 		{
 			name: "error getting users from store",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					ListFn: func(context.Context, meta.ListOptions) (UserList, error) {
 						return UserList{}, errors.New("error listing users")
 					},
@@ -49,7 +49,7 @@ func TestUserServiceList(t *testing.T) {
 		{
 			name: "success",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					ListFn: func(context.Context, meta.ListOptions) (UserList, error) {
 						return UserList{}, nil
 					},
@@ -83,7 +83,7 @@ func TestUsersServiceGet(t *testing.T) {
 		{
 			name: "with error from store",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					GetFn: func(context.Context, string) (User, error) {
 						return User{}, &meta.ErrNotFound{}
 					},
@@ -97,7 +97,7 @@ func TestUsersServiceGet(t *testing.T) {
 		{
 			name: "success",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					GetFn: func(context.Context, string) (User, error) {
 						return testUser, nil
 					},
@@ -126,7 +126,7 @@ func TestUsersServiceLock(t *testing.T) {
 		{
 			name: "error updating user in store",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					LockFn: func(context.Context, string) error {
 						return errors.New("store error")
 					},
@@ -141,7 +141,7 @@ func TestUsersServiceLock(t *testing.T) {
 		{
 			name: "error deleting user sessions from store",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					LockFn: func(context.Context, string) error {
 						return nil
 					},
@@ -162,7 +162,7 @@ func TestUsersServiceLock(t *testing.T) {
 		{
 			name: "success",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					LockFn: func(context.Context, string) error {
 						return nil
 					},
@@ -196,7 +196,7 @@ func TestUsersServiceUnlock(t *testing.T) {
 		{
 			name: "error updating user in store",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					UnlockFn: func(context.Context, string) error {
 						return errors.New("store error")
 					},
@@ -211,7 +211,7 @@ func TestUsersServiceUnlock(t *testing.T) {
 		{
 			name: "success",
 			service: &usersService{
-				usersStore: &mockUsersStore{
+				usersStore: &MockUsersStore{
 					UnlockFn: func(context.Context, string) error {
 						return nil
 					},
@@ -231,35 +231,4 @@ func TestUsersServiceUnlock(t *testing.T) {
 			testCase.assertions(err)
 		})
 	}
-}
-
-type mockUsersStore struct {
-	CreateFn func(context.Context, User) error
-	ListFn   func(context.Context, meta.ListOptions) (UserList, error)
-	GetFn    func(context.Context, string) (User, error)
-	LockFn   func(context.Context, string) error
-	UnlockFn func(context.Context, string) error
-}
-
-func (m *mockUsersStore) Create(ctx context.Context, user User) error {
-	return m.CreateFn(ctx, user)
-}
-
-func (m *mockUsersStore) List(
-	ctx context.Context,
-	opts meta.ListOptions,
-) (UserList, error) {
-	return m.ListFn(ctx, opts)
-}
-
-func (m *mockUsersStore) Get(ctx context.Context, id string) (User, error) {
-	return m.GetFn(ctx, id)
-}
-
-func (m *mockUsersStore) Lock(ctx context.Context, id string) error {
-	return m.LockFn(ctx, id)
-}
-
-func (m *mockUsersStore) Unlock(ctx context.Context, id string) error {
-	return m.UnlockFn(ctx, id)
 }
